ielts-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper
pydub
numpy
audioop-lts